Refactor languages utils
