+++ client/components/locale-suggestions/list-item.jsx
@@ -1,7 +1,7 @@
+import { getLanguage } from '@automattic/languages';
 import { getLocaleSlug } from 'i18n-calypso';
 import PropTypes from 'prop-types';
 import { Component } from 'react';
-import { getLanguage } from 'calypso/lib/i18n-utils';
 
 class LocaleSuggestionsListItem extends Component {
 	static propTypes = {
