+++ client/layout/masterbar/logged-out.jsx
@@ -7,7 +7,8 @@ import { withCurrentRoute } from 'calypso/components/route';
 import WordPressLogo from 'calypso/components/wordpress-logo';
 import WordPressWordmark from 'calypso/components/wordpress-wordmark';
 import { isDomainConnectAuthorizePath } from 'calypso/lib/domains/utils';
-import { isDefaultLocale, addLocaleToPath } from 'calypso/lib/i18n-utils';
+import { isDefaultLocale } from 'calypso/lib/i18n-utils';
+import { addLocaleToPath } from 'calypso/lib/i18n-utils/path';
 import { login } from 'calypso/lib/paths';
 import { addQueryArgs } from 'calypso/lib/route';
 import Item from './item';
