+++ client/lib/i18n-utils/node.js
@@ -7,22 +7,11 @@ import config from '@automattic/calypso-config';
 //
 // export * from './utils';
 export {
-	addLocaleToPath,
-	getLanguage,
-	getLanguageRouteParam,
-	getLanguageSlugs,
-	getLocaleFromPath,
 	isDefaultLocale,
-	isLocaleVariant,
-	isLocaleRtl,
 	localizeUrl,
 	canBeTranslated,
-	removeLocaleFromPath,
-	getPathParts,
-	filterLanguageRevisions,
 	translationExists,
 	isMagnificentLocale,
-	isTranslatedIncompletely,
 } from './utils';
 
 export const getLocaleSlug = () => config( 'i18n_default_locale_slug' );
