+++ client/me/account/main.jsx
@@ -1,6 +1,6 @@
 import config from '@automattic/calypso-config';
 import { Card, Button } from '@automattic/components';
-import languages from '@automattic/languages';
+import languages, { getLanguage } from '@automattic/languages';
 import debugFactory from 'debug';
 import emailValidator from 'email-validator';
 import { localize } from 'i18n-calypso';
@@ -34,7 +34,7 @@ import SitesDropdown from 'calypso/components/sites-dropdown';
 import PageViewTracker from 'calypso/lib/analytics/page-view-tracker';
 import { type as domainTypes } from 'calypso/lib/domains/constants';
 import { supportsCssCustomProperties } from 'calypso/lib/feature-detection';
-import { getLanguage, isLocaleVariant, canBeTranslated } from 'calypso/lib/i18n-utils';
+import { canBeTranslated } from 'calypso/lib/i18n-utils';
 import { ENABLE_TRANSLATOR_KEY } from 'calypso/lib/i18n-utils/constants';
 import { protectForm } from 'calypso/lib/protect-form';
 import twoStepAuthorization from 'calypso/lib/two-step-authorization';
@@ -187,7 +187,7 @@ class Account extends Component {
 			);
 		}
 
-		const localeVariantSelected = isLocaleVariant( value ) ? value : '';
+		const localeVariantSelected = getLanguage( value ).parentLangSlug ? value : '';
 
 		const originalSlug =
 			this.getUserSetting( 'locale_variant' ) || this.getUserSetting( 'language' ) || '';
