+++ client/my-sites/domains/domain-management/domain-connect/controller.jsx
@@ -1,4 +1,4 @@
-import { getLanguage } from 'calypso/lib/i18n-utils';
+import { getLanguage } from '@automattic/languages';
 import DomainConnectAuthorize from './domain-connect-authorize';
 import DomainConnectNotFoundError from './domain-connect-not-found-error';
