+++ client/my-sites/invites/controller.js
@@ -3,7 +3,7 @@ import i18n from 'i18n-calypso';
 import page from 'page';
 import { createElement } from 'react';
 import store from 'store';
-import { getLocaleFromPath, removeLocaleFromPath } from 'calypso/lib/i18n-utils';
+import { getLocaleFromPath, removeLocaleFromPath } from 'calypso/lib/i18n-utils/path';
 import { navigate } from 'calypso/lib/navigate';
 import InviteAccept from 'calypso/my-sites/invites/invite-accept';
 import { getRedirectAfterAccept } from 'calypso/my-sites/invites/utils';
