+++ client/my-sites/invites/index.js
@@ -1,6 +1,6 @@
 import page from 'page';
 import { makeLayout, render as clientRender } from 'calypso/controller';
-import { getLanguageRouteParam } from 'calypso/lib/i18n-utils';
+import { getLanguageRouteParam } from 'calypso/lib/i18n-utils/path';
 import { acceptInvite, redirectWithoutLocaleifLoggedIn } from './controller';
 
 export default () => {
