+++ client/my-sites/theme/index.node.js
@@ -1,6 +1,6 @@
 import { makeLayout, ssrSetupLocale } from 'calypso/controller';
 import { setHrefLangLinks, setLocalizedCanonicalUrl } from 'calypso/controller/localized-links';
-import { getLanguageRouteParam } from 'calypso/lib/i18n-utils';
+import { getLanguageRouteParam } from 'calypso/lib/i18n-utils/path';
 import { details, fetchThemeDetailsData, notFoundError } from './controller';
 
 export default function ( router ) {
